Find Roast Hit

Find the hit using the Roast method

find_roast_hit(
  screenR_Object,
  matrix_model,
  contrast,
  nrot = 9999,
  number_barcode = 3,
  direction = "Down",
  p_val = 0.05
)

Arguments

screenR_Object

The ScreenR object obtained using the create_screenR_object

matrix_model

The matrix that will be used to perform the linear model analysis. Created using model.matrix

contrast

A vector or a single value indicating the index or the name of the column the model_matrix to which perform the analysis

nrot

Number of rotation to perform the test

number_barcode

The hit find with the Roast method

direction

Direction of variation

p_val

P value cut off

Value

The hits found by ROAST method

Examples

obj <- get0('obj', envir = asNamespace('ScreenR'))
matrix_model <- model.matrix(~slot(obj, 'groups'))
colnames(matrix_model) <- c('Control', 'T0_T48', 'Treated')

find_roast_hit(obj, matrix_model = matrix_model, contrast = 'Treated')
#> Warning: 3rows with all zero counts
#> # A tibble: 22 × 9
#>    Gene   NGenes PropDown PropUp Direction PValue   FDR PValue.Mixed FDR.Mixed
#>    <chr>   <int>    <dbl>  <dbl> <fct>      <dbl> <dbl>        <dbl>     <dbl>
#>  1 HERC3      10    0.7      0   Down      0.0008 0.400       0.0007     0.347
#>  2 ACOX2      10    0.4      0   Down      0.0023 0.536       0.0156     0.738
#>  3 TRIM15     10    0.2      0   Down      0.0053 0.536       0.390      0.843
#>  4 PCMT1      10    0.4      0.1 Down      0.0078 0.536       0.0025     0.436
#>  5 TKTL1      10    0.1      0   Down      0.0092 0.536       0.316      0.843
#>  6 SDHB       10    0.2      0   Down      0.0099 0.536       0.258      0.843
#>  7 CHST13     10    0.4      0   Down      0.0108 0.536       0.0048     0.507
#>  8 EP300      10    0.4      0   Down      0.0122 0.536       0.002      0.436
#>  9 TRIM71      9    0.111    0   Down      0.0147 0.536       0.379      0.843
#> 10 SIN3B      10    0.3      0   Down      0.0175 0.582       0.310      0.843
#> # … with 12 more rows