Filter using the slope filter

This function is used to improve the quality of the hits found. It computes a regression line in the different samples ad uses the slope of this line to see the trend

Usage

filter_by_slope(
  screenR_Object,
  genes,
  group_var_treatment,
  group_var_control,
  slope_control,
  slope_treatment
)

Arguments

screenR_Object

The ScreenR object obtained using the create_screenr_object

genes

The genes for which the slope as to be computed. Those genes are the result of the three statistical methods selection

group_var_treatment

The variable to use as independent variable (x) for the linear model of the treatment

group_var_control

The variable to use as independent variable (x) for the linear model of the the control

slope_control

A value used as threshold for the control slope

slope_treatment

A value used as threshold for the treatment slope

Value

A data frame with the slope for the treatment and the control for each gene

Examples

object <- get0("object", envir = asNamespace("ScreenR"))

filter_by_slope(
    screenR_Object = object, genes = c("Gene_1", "Gene_2"),
    group_var_treatment = c("T1", "T2", "TRT"),
    group_var_control = c("T1", "T2", "Time3", "Time4"),
    slope_control = 0.5, slope_treatment = 1
)
#> # A tibble: 280 × 11
#>    Barcode Gene   Sample   Frequ…¹ Seque…² Library Gene_ID Day   Treat…³ slope…⁴
#>    <fct>   <chr>  <fct>      <dbl> <lgl>   <lgl>   <lgl>   <chr> <chr>     <dbl>
#>  1 10004   Gene_1 T1          54.7 NA      NA      NA      T1    T1         2.07
#>  2 10004   Gene_1 T2          33.5 NA      NA      NA      T2    T2         2.07
#>  3 10004   Gene_1 Time3_T…    22.9 NA      NA      NA      Time3 TRT        2.07
#>  4 10004   Gene_1 Time3_T…    45.2 NA      NA      NA      Time3 TRT        2.07
#>  5 10004   Gene_1 Time3_T…    18.6 NA      NA      NA      Time3 TRT        2.07
#>  6 10004   Gene_1 Time3_A     27.8 NA      NA      NA      Time3 Time3      2.07
#>  7 10004   Gene_1 Time3_B     50.1 NA      NA      NA      Time3 Time3      2.07
#>  8 10004   Gene_1 Time3_C     44.6 NA      NA      NA      Time3 Time3      2.07
#>  9 10004   Gene_1 Time4_T…    33.4 NA      NA      NA      Time4 TRT        2.07
#> 10 10004   Gene_1 Time4_T…    49.2 NA      NA      NA      Time4 TRT        2.07
#> # … with 270 more rows, 1 more variable: slope_control <dbl>, and abbreviated
#> #   variable names ¹​Frequency, ²​Sequence, ³​Treatment, ⁴​slope_treatment