Find Roast Hit

Find the hit using the Roast method

Usage

find_roast_hit(
  screenR_Object,
  matrix_model,
  contrast,
  nrot = 9999,
  number_barcode = 3,
  direction = "Down",
  p_val = 0.05
)

Arguments

screenR_Object

The ScreenR object obtained using the create_screenR_object

matrix_model

The matrix that will be used to perform the linear model analysis. Created using model.matrix

contrast

A vector or a single value indicating the index or the name of the column the model_matrix to which perform the analysis

nrot

Number of rotation to perform the test

number_barcode

The hit find with the Roast method

direction

Direction of variation

p_val

P value cut off

Value

The hits found by ROAST method

Examples

object <- get0("object", envir = asNamespace("ScreenR"))
matrix_model <- model.matrix(~ slot(object, "groups"))
colnames(matrix_model) <- c("Control", "T1_T2", "Treated")

result <- find_roast_hit(object,
    matrix_model = matrix_model,
    contrast = "Treated", nrot = 100
)
#> Warning: 3rows with all zero counts
head(result)
#> # A tibble: 6 × 9
#>   Gene     NGenes PropDown PropUp Direction  PValue   FDR PValue.Mixed FDR.Mixed
#>   <chr>     <int>    <dbl>  <dbl> <fct>       <dbl> <dbl>        <dbl>     <dbl>
#> 1 Gene_173     10    0.7      0   Down      0.00990 0.330      0.00990     0.529
#> 2 Gene_121     10    0.4      0   Down      0.00990 0.330      0.00990     0.529
#> 3 Gene_293     10    0.4      0.1 Down      0.00990 0.330      0.0396      0.661
#> 4 Gene_377     10    0.2      0   Down      0.00990 0.330      0.208       0.856
#> 5 Gene_457     10    0.2      0   Down      0.00990 0.330      0.376       0.856
#> 6 Gene_502      9    0.111    0   Down      0.00990 0.330      0.406       0.856